+++ src/main/java/com/demo/demo/dto/InProgressCourseResponse.java
@@ -2,10 +2,11 @@ package com.demo.demo.dto;
 
 import com.demo.demo.enums.AgeGroup;
 import lombok.Data;
+
 import java.time.LocalDateTime;
 
 @Data
-public class CourseResponse {
+public class InProgressCourseResponse {
     private Long id;
     private String title;
     private String description;
@@ -13,5 +14,7 @@ public class CourseResponse {
     private String overview;
     private AgeGroup ageGroup;
     private LocalDateTime createdAt;
-    private long totalEnrollment;
+    private long totalChapter;
+    private long completedChapter;
 }
+
